Moved LinkWalker into its own file
